Windows compilation works.  Ready to ship 3.10ALPHA1
git-svn-id: https://svn.nmap.org/trunk/nmap@1508 e0a8ed71-7df4-0310-8962-fdc924857419

+++ tcpip.cc
@@ -121,7 +121,7 @@ int resolve(char *hostname, struct sockaddr_storage *ss, size_t *sslen,
   rc = getaddrinfo(hostname, NULL, &hints, &result);
   if (rc != 0)
     return 0;
-  assert(result->ai_addrlen > 0 && result->ai_addrlen <= sizeof(struct sockaddr_storage));
+  assert(result->ai_addrlen > 0 && result->ai_addrlen <= (int) sizeof(struct sockaddr_storage));
   *sslen = result->ai_addrlen;
   memcpy(ss, result->ai_addr, *sslen);
   freeaddrinfo(result);
